move dataclass_repr function to a new display module
